I'm not gonna get hung up the math but trying to figure out what numbers correspond to what in the instructions is wildly confusing 11 2 32 43		69  82
5 6
			461 436
7 8

there's no n length combination of 11, 2, 32 and 43 that can multiply out to 69 or 82

even brining in the second square that is left at default values -- still we can not achieve these numbers by multiplication along. If I'm missing something here please help me.

so whatever, the operand Matrix*int is going to return the total sum of the Matrix times an Int -- If this becomes an issues I will happily switch it to any particular sub units of the n sized matrix times an int -- additionally we could switch this to an infix, but again that will require some more understanding