ok cool all the overloads and infixs and whatever garbage is done
The requirements for this project are as follows:

Create a custom Matrix class which will.
check

Store matrixes in a 2-dimentional array
I think this is bad english, I take this to mean Matrix classes should contain at least a 2 dimensional array

Overload the * and index operators
done, again with caviat that multiplication is perhaps not as the designer of this exercise intended although it could be made to be whatever trivially as the current solution traverses the 2 dimensional array and could pick whatever bits it aliked along the way.

Include height and width properties with setters that will resize your matrix data array when they're changed
check